[Agent] Fix panic at unreachable code #19895

+++ agent/src/dispatcher/recv_engine/bpf.rs
@@ -109,9 +109,9 @@ impl Builder {
     fn skip_ethernet(&self) -> BpfBuilder {
         let mut bpf_builder = BpfBuilder::default();
         let eth_type = if self.is_ipv6 {
-            EthernetType::Ipv6 as u32
+            u16::from(EthernetType::Ipv6) as u32
         } else {
-            EthernetType::Ipv4 as u32
+            u16::from(EthernetType::Ipv4) as u32
         };
 
         bpf_builder
@@ -121,7 +121,7 @@ impl Builder {
             }))
             .append(BpfSyntax::JumpIf(JumpIf {
                 cond: JumpTest::JumpNotEqual,
-                val: EthernetType::Dot1Q as u32,
+                val: u16::from(EthernetType::Dot1Q) as u32,
                 skip_true: 2,
                 skip_false: 0,
             }))
@@ -173,7 +173,7 @@ impl Builder {
             }))
             .append(BpfSyntax::JumpIf(JumpIf {
                 cond: JumpTest::JumpNotEqual,
-                val: IpProtocol::Udp as u32,
+                val: u8::from(IpProtocol::Udp) as u32,
                 skip_true: 4,
                 ..Default::default()
             }))
@@ -200,7 +200,7 @@ impl Builder {
             .branch(
                 JumpIf {
                     cond: JumpTest::JumpNotEqual,
-                    val: IpProtocol::Gre as u32,
+                    val: u8::from(IpProtocol::Gre) as u32,
                     ..Default::default()
                 },
                 Self::bypass_modifier,
@@ -232,7 +232,7 @@ impl Builder {
             }))
             .append(BpfSyntax::JumpIf(JumpIf {
                 cond: JumpTest::JumpNotEqual,
-                val: IpProtocol::Udp as u32,
+                val: u8::from(IpProtocol::Udp) as u32,
                 skip_true: 4,
                 ..Default::default()
             }))
@@ -259,7 +259,7 @@ impl Builder {
             .branch(
                 JumpIf {
                     cond: JumpTest::JumpNotEqual,
-                    val: IpProtocol::Gre as u32,
+                    val: u8::from(IpProtocol::Gre) as u32,
                     ..Default::default()
                 },
                 Self::bypass_modifier,
@@ -304,14 +304,14 @@ impl Builder {
             }))
             .append(BpfSyntax::JumpIf(JumpIf {
                 cond: JumpTest::JumpNotEqual,
-                val: IpProtocol::Tcp as u32,
+                val: u8::from(IpProtocol::Tcp) as u32,
                 skip_false: 1,
                 ..Default::default()
             }))
             .branch(
                 JumpIf {
                     cond: JumpTest::JumpNotEqual,
-                    val: IpProtocol::Udp as u32,
+                    val: u8::from(IpProtocol::Udp) as u32,
                     ..Default::default()
                 },
                 Self::bypass_modifier,
@@ -408,7 +408,7 @@ impl Builder {
             }))
             .append(BpfSyntax::JumpIf(JumpIf {
                 cond: JumpTest::JumpNotEqual,
-                val: IpProtocol::Tcp as u32,
+                val: u8::from(IpProtocol::Tcp) as u32,
                 skip_true: 2,
                 ..Default::default()
             }))
@@ -425,7 +425,7 @@ impl Builder {
             .branch(
                 JumpIf {
                     cond: JumpTest::JumpNotEqual,
-                    val: IpProtocol::Udp as u32,
+                    val: u8::from(IpProtocol::Udp) as u32,
                     ..Default::default()
                 },
                 Self::bypass_modifier,
@@ -469,7 +469,7 @@ impl Builder {
             }))
             .append(BpfSyntax::JumpIf(JumpIf {
                 cond: JumpTest::JumpNotEqual,
-                val: IpProtocol::Tcp as u32,
+                val: u8::from(IpProtocol::Tcp) as u32,
                 skip_true: 2,
                 ..Default::default()
             }))
@@ -486,7 +486,7 @@ impl Builder {
             .branch(
                 JumpIf {
                     cond: JumpTest::JumpNotEqual,
-                    val: IpProtocol::Udp as u32,
+                    val: u8::from(IpProtocol::Udp) as u32,
                     ..Default::default()
                 },
                 Self::bypass_modifier,
@@ -537,7 +537,7 @@ impl Builder {
             }))
             .append(BpfSyntax::JumpIf(JumpIf {
                 cond: JumpTest::JumpNotEqual,
-                val: IpProtocol::Tcp as u32,
+                val: u8::from(IpProtocol::Tcp) as u32,
                 skip_true: 2,
                 ..Default::default()
             }))
@@ -554,7 +554,7 @@ impl Builder {
             .branch(
                 JumpIf {
                     cond: JumpTest::JumpNotEqual,
-                    val: IpProtocol::Udp as u32,
+                    val: u8::from(IpProtocol::Udp) as u32,
                     ..Default::default()
                 },
                 Self::bypass_modifier,
@@ -594,7 +594,7 @@ impl Builder {
             }))
             .append(BpfSyntax::JumpIf(JumpIf {
                 cond: JumpTest::JumpNotEqual,
-                val: IpProtocol::Tcp as u32,
+                val: u8::from(IpProtocol::Tcp) as u32,
                 skip_true: 2,
                 ..Default::default()
             }))
@@ -611,7 +611,7 @@ impl Builder {
             .branch(
                 JumpIf {
                     cond: JumpTest::JumpNotEqual,
-                    val: IpProtocol::Udp as u32,
+                    val: u8::from(IpProtocol::Udp) as u32,
                     ..Default::default()
                 },
                 Self::bypass_modifier,
